+++ src/main/scala/EnsimeCoursierPlugin.scala
@@ -4,13 +4,11 @@ package org.ensime
 
 import scala.util.Properties.{versionNumberString => sbtScalaVersion}
 
+import CoursierHelper._
 import EnsimeCoursierKeys._
 import EnsimeKeys._
 import sbt._
-import sbt.IO._
 import sbt.Keys._
-import scalaz._
-import scalaz.concurrent.Task
 
 object EnsimeCoursierKeys {
   val ensimeServerVersion = settingKey[String](
@@ -24,7 +22,7 @@ object EnsimeCoursierKeys {
 
   def addEnsimeScalaPlugin(module: ModuleID, args: String = ""): Seq[Setting[_]] = {
     ensimeScalacOptions += {
-      val jar = EnsimeCoursierPlugin.resolveSingleJar(module, ensimeScalaVersion.value, ensimeRepositoryUrls.value)
+      val jar = resolveSingleJar(module, ensimeScalaVersion.value, ensimeRepositoryUrls.value)
       s"-Xplugin:${jar.getCanonicalFile}${args}"
     }
   }
@@ -39,7 +37,6 @@ object EnsimeCoursierKeys {
  * to avoid a dependency on coursier and provide hard coded jar paths.
  */
 object EnsimeCoursierPlugin extends AutoPlugin {
-
   override def requires = EnsimePlugin
   override def trigger = allRequirements
   val autoImport = EnsimeCoursierKeys
@@ -53,64 +50,10 @@ object EnsimeCoursierPlugin extends AutoPlugin {
       "https://oss.sonatype.org/content/repositories/snapshots/"
     ),
 
-    ensimeScalaJars := resolveScalaJars(scalaOrganization.value, ensimeScalaVersion.value)(resolvers(ensimeRepositoryUrls.value)),
-    ensimeScalaProjectJars := resolveScalaJars("org.scala-lang", sbtScalaVersion)(resolvers(ensimeRepositoryUrls.value)),
-    ensimeServerJars := resolveEnsimeJars(scalaOrganization.value, ensimeScalaVersion.value, ensimeServerVersion.value)(resolvers(ensimeRepositoryUrls.value)),
-    ensimeServerProjectJars := resolveEnsimeJars("org.scala-lang", sbtScalaVersion, ensimeServerVersion.value)(resolvers(ensimeRepositoryUrls.value))
+    ensimeScalaJars := resolveScalaJars(scalaOrganization.value, ensimeScalaVersion.value, ensimeRepositoryUrls.value),
+    ensimeScalaProjectJars := resolveScalaJars("org.scala-lang", sbtScalaVersion, ensimeRepositoryUrls.value),
+    ensimeServerJars := resolveEnsimeJars(scalaOrganization.value, ensimeScalaVersion.value, ensimeServerVersion.value, ensimeRepositoryUrls.value),
+    ensimeServerProjectJars := resolveEnsimeJars("org.scala-lang", sbtScalaVersion, ensimeServerVersion.value, ensimeRepositoryUrls.value)
   )
 
-  private[this] def resolvers(urls: Seq[String]) = urls.map(coursier.MavenRepository(_))
-
-  private[this] def resolve(modules: ModuleID*)(implicit repos: Seq[coursier.Repository]): Seq[File] = {
-    val resolution = coursier.Resolution(
-      modules.map { module =>
-      coursier.Dependency(
-        coursier.Module(
-          module.organization,
-          module.name
-        ),
-        module.revision,
-        configuration = module.configurations.getOrElse(""),
-        transitive = module.isTransitive
-      )
-    }.toSet
-    )
-
-    val fetch = coursier.Fetch.from(repos, coursier.Cache.fetch())
-    val resolved = resolution.process.run(fetch).unsafePerformSync
-    resolved.errors.foreach { err =>
-      throw new RuntimeException(s"failed to resolve $err")
-    }
-
-    Task.gatherUnordered(
-      resolved.artifacts.map(coursier.Cache.file(_).run)
-    ).unsafePerformSync.flatMap {
-        case -\/(err)                                    => throw new RuntimeException(err.message)
-        case \/-(file) if !file.getName.endsWith(".jar") => None
-        case \/-(file)                                   => Some(file)
-      }
-  }
-
-  private[this] def resolveScalaJars(org: String, version: String)(implicit repos: Seq[coursier.Repository]): Seq[File] = resolve(
-    org % "scalap" % version intransitive,
-    org % "scala-compiler" % version intransitive,
-    org % "scala-reflect" % version intransitive,
-    org % "scala-library" % version intransitive
-  )
-
-  private[this] def resolveEnsimeJars(org: String, scala: String, ensime: String)(implicit repos: Seq[coursier.Repository]): Seq[File] = {
-    val Some((major, minor)) = CrossVersion.partialVersion(scala)
-    resolve(
-      "org.ensime" % s"server_$major.$minor" % ensime,
-      org % "scalap" % scala intransitive
-    )
-  }
-
-  private[ensime] def resolveSingleJar(module: ModuleID, scalaVersion: String, repositories: Seq[String]): File = {
-    val mod = CrossVersion(
-      scalaVersion,
-      CrossVersion.binaryScalaVersion(scalaVersion)
-    )(module.intransitive)
-    resolve(mod)(resolvers(repositories)).head.getCanonicalFile
-  }
 }
